Client/requirements.txt
asgiref==3.6.0
blinker==1.6.2
click==8.1.3
Flask==2.3.2
grpcio==1.54.0
grpcio-tools==1.54.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
protobuf==4.22.4
Werkzeug==2.3.3
